update 1-concurrent_coroutines.py file

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -10,7 +10,7 @@ from typing import List
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: int) -> List[int]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     """
     Spawns n number of random delay with max_delay been the highest
     possible delay.
